Phase 1 completed
