--- freelancer_scraper/freelancer_scraper/db.py
@@ -1,3 +1,0 @@
-from pymongo import MongoClient
-
-mongo_client = MongoClient("mongodb://localhost:27017")
