--- freelancer_scraper/scrapy.cfg
@@ -1,11 +1,0 @@
-# Automatically created by: scrapy startproject
-#
-# For more information about the [deploy] section see:
-# https://scrapyd.readthedocs.io/en/latest/deploy.html
-
-[settings]
-default = freelancer_scraper.settings
-
-[deploy]
-#url = http://localhost:6800/
-project = freelancer_scraper
